+++ WolvenKit.Common/IWitcherArchiveType.cs
