+++ WolvenKit.Common/IWitcherFile.cs
