+++ WolvenKit.Common/Properties/AssemblyInfo.cs
