+++ WolvenKit.Common/WitcherTreeNode.cs
